Completed the profile settings page
I added the button from settings into the profile page so that the settings page for a user profile can be accessed quickly. I also completed the profile settings page so that users can set location or request to become a CrisisRelief worker. Finally, I adjusted the logo so that it wasn't as tall.