Completed .env transition, foodbanks, and proper docker implementation
When using docker, go to to the application folder and run 'docker-compose up --build'. It is important to run 'npm run build' in the backend directory first due to a weird interaction. We had to switch to bcryptjs because regular bcrypt doesn't work with the what the docker runs on. Regular bcrypt works if you're on windows I believe.